fixed a bug in populating posts

+++ controllers/postController.js
@@ -38,16 +38,18 @@ module.exports = {
     const postId = req.params.id;
     Post.findByIdAndRemove(postId)
       .then((post) => {
+        if (post.imageUrl != null) {
+          let imageID = post.imageUrl.substring(11, post.imageUrl.length);
+  
+           fs.unlink(`./public/uploads/${imageID}`, (err) => {
+            if (err) throw err;
+            req.flash("success", `Post deleted successfully!`);
+            res.locals.redirect = `/profile/${req.user._id}`;
+            next();
+            console.log('successfully deleted /tmp/hello');
+          });
 
-        let imageID = post.imageUrl.substring(11, post.imageUrl.length);
-
-         fs.unlink(`./public/uploads/${imageID}`, (err) => {
-          if (err) throw err;
-          req.flash("success", `Post deleted successfully!`);
-          res.locals.redirect = `/profile/${req.user._id}`;
-          next();
-          console.log('successfully deleted /tmp/hello');
-        });
+        }
 
       })
       .catch(error => {
